+++ src/main/java/org/levelup/bank/system/repository/AccountRepository.java
@@ -8,9 +8,9 @@ import java.util.Collection;
 public interface AccountRepository {
 
     @Timed
-    Collection<Account> findUserAccounts(long userId);
+    Collection<Account> findUserAccounts(int userId);
 
-    Account createAccount(String number, String currency, long clientId);
+    Account createAccount(String number, String currency, int clientId);
     Account editAccountNumber(long accountId, String number);
     Boolean deleteAccount(long accountId, long clientId);
